requirements.txt
aiohttp==3.4.4
altgraph==0.16.1
amqp==2.3.2
anyjson==0.3.3
APScheduler==3.5.3
argcomplete==1.8.1
arrow==0.12.1
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.0
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
bcrypt==3.1.4
beautifulsoup4==4.7.1
billiard==3.5.0.4
blessed==1.15.0
cairocffi==1.0.2
CairoSVG==2.4.0
celery==4.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
chartjs==1.2
Click==7.0
conda==4.3.16
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
croniter==0.3.25
cryptography==2.3.1
cssselect==1.0.3
cssselect2==0.2.1
decorator==4.3.0
dj-database-url==0.5.0
Django==2.2.1
django-bower==5.2.0
django-celery==3.2.2
django-celery-beat==1.2.0
django-chartjs==1.3
django-chronograph-ispm==0.0.3
django-cron==0.5.1
django-dbbackup==3.2.0
django-extensions==2.1.3
django-future==0.2.3
django-google-charts==0.1.1
django-grappelli==2.11.1
django-heroku==0.3.1
django-model-utils==3.1.2
django-periodically==0.3.0
django-picklefield==1.1.0
django-q==1.0.1
django-qsstats-magic==1.0.0
django-recaptcha==1.4.0
django-rest-swagger==2.2.0
django-rq==1.2.0
django-rq-jobs==0.1.5.1
django-rq-scheduler==1.1.1
django-scheduler==0.8.8
django-selenium==0.9.8
django-ssl-auth==2.0.1
django-sslserver==0.20
django-timezone-field==3.0
djangorestframework==3.9.4
entrypoints==0.2.3
Events==0.3
Flask==1.0.2
future==0.17.0
geopy==1.17.0
glamkit-eventtools==0.5.1
google==2.0.1
h5py==2.8.0
html5lib==1.0.1
hupper==1.3
hyperlink==18.0.0
icalendar==4.0.3
incremental==17.5.0
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10.1
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
List==1.3.0
lxml==4.2.5
multidict==4.4.1
netifaces==0.10.4
numpy==1.15.3
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
openapi-codec==1.3.2
paramiko==2.4.2
parsel==1.5.0
passlib==1.7.1
PasteDeploy==1.5.2
pbr==4.2.0
pdfrw==0.4
pefile==2018.8.8
pexpect==4.2.1
Pillow==6.0.0
plaster==1.0
plaster-pastedeploy==0.6
pluggy==0.6.0
psutil==5.4.8
psycopg2==2.8.2
psycopg2-binary==2.7.7
py==1.5.2
py2exe==0.9.2.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pygame==1.9.4
pygooglechart==0.4.0
PyHamcrest==1.9.0
PyInstaller==3.4
pylint==2.1.1
pyOpenSSL==18.0.0
PyPDF2==1.26.0
Pyphen==0.9.5
pyramid==1.9.2
python-crontab==2.3.5
python-dateutil==2.8.0
python-debian==0.1.32
python-gnupg==0.4.1
pytz==2019.1
pyxdg==0.25
PyYAML==3.13
ratelim==0.1.6
redis==2.10.6
reportlab==3.4.0
repoze.lru==0.7
requests==2.22.0
requests-unixsocket==0.1.5
rq==0.12.0
rq-scheduler==0.8.3
ruamel.yaml==0.15.88
schedule==0.5.0
scipy==1.1.0
Scrapy==1.5.1
SecretStorage==3.1.0
service-identity==17.0.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
sqlparse==0.3.0
testresources==2.0.1
tinycss2==1.0.2
tox==2.5.0
translationstring==1.3
Twisted==18.9.0
typed-ast==1.1.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.25.2
virtualenv==15.1.0
WeasyPrint==47
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.10.11
yarl==1.2.6
gunicorn
google-api-python-client
markdown
django-filter
setuptools
ordereddict
django-cors-headers